+++ CHANGELOG.md
@@ -25,6 +25,7 @@ If you like this project and find it useful, please consider giving it a star on
 ### Changed
 
 - [package]: Update dependencies.
+- [spawn]: Modified the install / update function to add more info in the log.
 
 <a href="https://www.buymeacoffee.com/luligugithub">
   <img src="./yellow-button.png" alt="Buy me a coffee" width="120">